+++ src/private/release/Get-UpdatedChangelogReferenceFooter.ps1
@@ -30,20 +30,28 @@ function Get-ChangelogReleaseLink {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory)]
-        [string]$ReleaseTagPrefix,
-        [Parameter(Mandatory)]
-        [string]$UnreleasedCompareLinkPrefix,
+        [pscustomobject]$Context,
         [AllowEmptyString()]
         [string]$PreviousReleaseReference,
         [Parameter(Mandatory)]
-        [string]$ReleaseTag
+        [string]$ReleaseReference
     )
 
     if ([string]::IsNullOrWhiteSpace($PreviousReleaseReference)) {
-        return "$ReleaseTagPrefix$ReleaseTag"
+        if (-not [string]::IsNullOrWhiteSpace($Context.ReleaseTagPrefix)) {
+            return "$($Context.ReleaseTagPrefix)$ReleaseReference"
+        }
+
+        return Get-KeepAChangelogCompareLink `
+            -RepositoryLinkData $Context `
+            -BaseReference $ReleaseReference `
+            -TargetReference $ReleaseReference
     }
 
-    return "$UnreleasedCompareLinkPrefix$PreviousReleaseReference...$ReleaseTag"
+    return Get-KeepAChangelogCompareLink `
+        -RepositoryLinkData $Context `
+        -BaseReference $PreviousReleaseReference `
+        -TargetReference $ReleaseReference
 }
 
 function Add-ChangelogReferenceLabel {
@@ -100,12 +108,14 @@ function Get-UpdatedChangelogReferenceFooter {
         $orderedLabelList.Add('Unreleased')
     }
 
-    $updatedUnreleasedLink = "$($Context.UnreleasedCompareLinkPrefix)$($Release.Tag)...HEAD"
+    $updatedUnreleasedLink = Get-KeepAChangelogCompareLink `
+        -RepositoryLinkData $Context `
+        -BaseReference $Release.Reference `
+        -TargetReference $Context.UnreleasedTargetReference
     $newReleaseLink = Get-ChangelogReleaseLink `
-        -ReleaseTagPrefix $Context.ReleaseTagPrefix `
-        -UnreleasedCompareLinkPrefix $Context.UnreleasedCompareLinkPrefix `
+        -Context $Context `
         -PreviousReleaseReference $Context.PreviousReleaseReference `
-        -ReleaseTag $Release.Tag
+        -ReleaseReference $Release.Reference
     $linkMap['Unreleased'] = $updatedUnreleasedLink
 
     Add-ChangelogReferenceLabel -OrderedLabelList $orderedLabelList -LinkMap $linkMap -Label $Release.Version